Move genetic value classes to separate source files. Note: we have renamed the Python classes here, which breaks API!
