+++ fwdpy11/src/genetic_values/init.cc
@@ -2,6 +2,11 @@
 
 namespace py = pybind11;
 
+void init_GeneticValue(py::module&);
+void init_GeneticValueWithMapping(py::module&);
+void init_Additive(py::module&);
+void init_Multiplicative(py::module&);
+void init_GBR(py::module&);
 void init_MultivariateGeneticValueToFitnessMap(py::module&);
 void init_MultivariateGSS(py::module&);
 void init_MultivariateGSSmo(py::module&);
@@ -11,6 +16,8 @@ void init_DiploidMultivariateEffectsStrictAdditive(py::module&);
 void
 init_base_classes(py::module& m)
 {
+    init_GeneticValue(m);
+    init_GeneticValueWithMapping(m);
     init_MultivariateGeneticValueToFitnessMap(m);
     init_DiploidPopulationMultivariateGeneticValueWithMapping(m);
 }
@@ -25,6 +32,9 @@ init_gvalue_to_fitness_classes(py::module& m)
 void
 init_genetic_value_classes(py::module& m)
 {
+    init_Additive(m);
+    init_Multiplicative(m);
+    init_GBR(m);
     init_DiploidMultivariateEffectsStrictAdditive(m);
 }
